rename: #17 뷰 이름 변경
